// src/app/models/consumer-vo.ts
export class ConsumerVo {
        _id: string; 
        account_id: string; 
        addr1: string; 
        addr2: string; 
        annul_electric_bill: string; 
        bill_type: string; 
        bn: string; 
        businessName: string; 
        c_email: string; 
        c_fn: string; 
        c_ln: string; 
        c_phone: string; 
        city: string; 
        create: string; 
        create_ts: number;
        crn: string; 
        current_annual_consumption: string; 
        last_quote_amount: string; 
        last_update: number;
        lastupdate: string; 
        leadURL: string; 
        next_action: string; 
        next_action_date: number;
        rateName: string; 
        source: string; 
        sp_email: string; 
        sp_fn: string; 
        sp_ln: string; 
        slo_fn: string; 
        slo_ln: string; 
        slo_email: string; 
        slo_fullname: string; 
        star_rating: number;
        state: string; 
        status: string; 
        type: string; 
        update_ts: string; 
        utilityName: string; 
        zip: string; 
        annual_electric_bill_i18n: string; 
        create_i18n: string; 
        lastupdate_i18n: string; 
        update_i18n: string; 
        current_annual_consumption_i18n: string; 
        last_quote_amount_i18n: string; 
        next_action_i18n: string;
        
        constructor() {
    
        }
     
}